Convert score module to package
