+++ spdivik/score/_optimizer.py
@@ -1,36 +1,12 @@
 from itertools import product
-from typing import Callable, List, NamedTuple, Tuple
 
 import numpy as np
-import pandas as pd
+from typing import Callable, NamedTuple, List, Tuple
 
-from spdivik.distance import DistanceMetric
-from spdivik.types import \
-    Centroids, \
-    Data, \
-    IntLabels, \
-    Quality, \
-    SegmentationMethod
-
+from spdivik.types import Data, IntLabels, Centroids, SegmentationMethod, \
+    Quality
 
 Score = Callable[[Data, IntLabels, Centroids], float]
-
-
-def dunn(data: Data, labels: IntLabels, centroids: Centroids,
-         distance: DistanceMetric) -> float:
-    if centroids.shape[0] == 1:
-        raise ValueError('At least 2 clusters are required.')
-    clusters = pd.DataFrame(data).groupby(labels).apply(lambda cluster: cluster.values)
-    intercluster = distance(centroids, centroids)
-    intercluster = np.min(intercluster[intercluster != 0])
-    intracluster = np.max([
-        np.mean(distance(cluster, centroid.reshape(1, -1)))
-        for cluster, centroid in zip(clusters, centroids)
-    ])
-    score = intercluster / intracluster
-    return score
-
-
 ParameterValues = NamedTuple('ParameterValues', [
     ('name', str),
     ('values', List)
