+++ test/score/optimizer.py
@@ -1,22 +1,10 @@
 import unittest
-from unittest.mock import MagicMock, patch
-
-from typing import List
+from unittest.mock import MagicMock
 
 import numpy as np
+from typing import List
 
-from spdivik.distance import KnownMetric, ScipyDistance
 import spdivik.score
-
-
-class TestDunn(unittest.TestCase):
-    def test_computes_inter_to_intracluster_distances_rate(self):
-        data = np.array([[1], [3], [4], [6]])
-        centroids = np.array([[2], [5]])
-        labels = np.array([1, 1, 2, 2], dtype=int)
-        distance = ScipyDistance(KnownMetric.euclidean)
-        dunn = spdivik.score.dunn(data, labels, centroids, distance)
-        self.assertAlmostEqual(dunn, 3.)
 
 
 def simulate_values(values: List):
